Rename imports of `roadmapper`
* Add init file to `src` folder
* Prefix imports of `roadmapper` with `src`
* Optimise imports/clean up unused imports
