+++ src/roadmapper/milestone.py
@@ -19,10 +19,10 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
+
 from datetime import datetime
 from dataclasses import dataclass, field
-from roadmapper.painter import Painter
-from roadmapper.timeline import Timeline
+from src.roadmapper.painter import Painter
 
 
 @dataclass(kw_only=True)
