+++ src/roadmapper/task.py
@@ -19,14 +19,13 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
-from datetime import datetime, date, timedelta
-from dateutil.relativedelta import relativedelta
-from dataclasses import dataclass, field
-from contextlib import contextmanager
 
-from roadmapper.painter import Painter
-from roadmapper.timeline import Timeline
-from roadmapper.milestone import Milestone
+from dataclasses import dataclass, field
+from datetime import datetime
+
+from src.roadmapper.milestone import Milestone
+from src.roadmapper.painter import Painter
+from src.roadmapper.timeline import Timeline
 
 
 @dataclass(kw_only=True)
